feedtracker/requirements.txt
mysqlclient==2.2.7
pymysql==1.1.0
python-decouple==3.8
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.3.0
gunicorn==23.0.0
whitenoise==6.8.2